requirements.txt
certifi==2020.6.20
pyjwt==1.7.1
requests==2.24.0